Core: Extract QUnit.module to its own file
As requested in https://github.com/qunitjs/qunit/pull/1168#discussion_r117645535
