+++ src/core.js
@@ -2,12 +2,13 @@ import { window, setTimeout } from "./globals";
 
 import equiv from "./equiv";
 import dump from "./dump";
+import module from "./module";
 import Assert from "./assert";
 import Test, { test, skip, only, todo, pushFailure } from "./test";
 import exportQUnit from "./export";
 
 import config from "./core/config";
-import { defined, extend, objectType, is, now, generateHash } from "./core/utilities";
+import { defined, extend, objectType, is, now } from "./core/utilities";
 import { registerLoggingCallbacks, runLoggingCallbacks } from "./core/logging";
 import { sourceFromStacktrace } from "./core/stacktrace";
 import ProcessingQueue from "./core/processing-queue";
@@ -18,7 +19,6 @@ import { on, emit } from "./events";
 import onError from "./core/onerror";
 import onUnhandledRejection from "./core/on-unhandled-rejection";
 
-let focused = false;
 const QUnit = {};
 export const globalSuite = new SuiteReport();
 
@@ -27,7 +27,6 @@ export const globalSuite = new SuiteReport();
 // it since each module has a suiteReport associated with it.
 config.currentModule.suiteReport = globalSuite;
 
-const moduleStack = [];
 var globalStartCalled = false;
 var runStarted = false;
 
@@ -36,140 +35,6 @@ QUnit.isLocal = !( defined.document && window.location.protocol !== "file:" );
 
 // Expose the current QUnit version
 QUnit.version = "@VERSION";
-
-function createModule( name, testEnvironment, modifiers ) {
-	const parentModule = moduleStack.length ? moduleStack.slice( -1 )[ 0 ] : null;
-	const moduleName = parentModule !== null ? [ parentModule.name, name ].join( " > " ) : name;
-	const parentSuite = parentModule ? parentModule.suiteReport : globalSuite;
-
-	const skip = parentModule !== null && parentModule.skip || modifiers.skip;
-	const todo = parentModule !== null && parentModule.todo || modifiers.todo;
-
-	const module = {
-		name: moduleName,
-		parentModule: parentModule,
-		tests: [],
-		moduleId: generateHash( moduleName ),
-		testsRun: 0,
-		unskippedTestsRun: 0,
-		childModules: [],
-		suiteReport: new SuiteReport( name, parentSuite ),
-
-		// Pass along `skip` and `todo` properties from parent module, in case
-		// there is one, to childs. And use own otherwise.
-		// This property will be used to mark own tests and tests of child suites
-		// as either `skipped` or `todo`.
-		skip: skip,
-		todo: skip ? false : todo
-	};
-
-	const env = {};
-	if ( parentModule ) {
-		parentModule.childModules.push( module );
-		extend( env, parentModule.testEnvironment );
-	}
-	extend( env, testEnvironment );
-	module.testEnvironment = env;
-
-	config.modules.push( module );
-	return module;
-}
-
-function processModule( name, options, executeNow, modifiers = {} ) {
-	let module = createModule( name, options, modifiers );
-
-	// Move any hooks to a 'hooks' object
-	const testEnvironment = module.testEnvironment;
-	const hooks = module.hooks = {};
-
-	setHookFromEnvironment( hooks, testEnvironment, "before" );
-	setHookFromEnvironment( hooks, testEnvironment, "beforeEach" );
-	setHookFromEnvironment( hooks, testEnvironment, "afterEach" );
-	setHookFromEnvironment( hooks, testEnvironment, "after" );
-
-	function setHookFromEnvironment( hooks, environment, name ) {
-		const potentialHook = environment[ name ];
-		hooks[ name ] = typeof potentialHook === "function" ? [ potentialHook ] : [];
-		delete environment[ name ];
-	}
-
-	const moduleFns = {
-		before: setHookFunction( module, "before" ),
-		beforeEach: setHookFunction( module, "beforeEach" ),
-		afterEach: setHookFunction( module, "afterEach" ),
-		after: setHookFunction( module, "after" )
-	};
-
-	const currentModule = config.currentModule;
-	if ( objectType( executeNow ) === "function" ) {
-		moduleStack.push( module );
-		config.currentModule = module;
-		executeNow.call( module.testEnvironment, moduleFns );
-		moduleStack.pop();
-		module = module.parentModule || currentModule;
-	}
-
-	config.currentModule = module;
-}
-
-// TODO: extract this to a new file alongside its related functions
-function module( name, options, executeNow ) {
-	if ( focused ) {
-		return;
-	}
-
-	if ( arguments.length === 2 ) {
-		if ( objectType( options ) === "function" ) {
-			executeNow = options;
-			options = undefined;
-		}
-	}
-
-	processModule( name, options, executeNow );
-}
-
-module.only = function() {
-	if ( focused ) {
-		return;
-	}
-
-	config.modules.length = 0;
-	config.queue.length = 0;
-
-	module( ...arguments );
-
-	focused = true;
-};
-
-module.skip = function( name, options, executeNow ) {
-	if ( focused ) {
-		return;
-	}
-
-	if ( arguments.length === 2 ) {
-		if ( objectType( options ) === "function" ) {
-			executeNow = options;
-			options = undefined;
-		}
-	}
-
-	processModule( name, options, executeNow, { skip: true } );
-};
-
-module.todo = function( name, options, executeNow ) {
-	if ( focused ) {
-		return;
-	}
-
-	if ( arguments.length === 2 ) {
-		if ( objectType( options ) === "function" ) {
-			executeNow = options;
-			options = undefined;
-		}
-	}
-
-	processModule( name, options, executeNow, { todo: true } );
-};
 
 extend( QUnit, {
 	on,
@@ -311,12 +176,6 @@ export function begin() {
 
 	config.blocking = false;
 	ProcessingQueue.advance();
-}
-
-function setHookFunction( module, hookName ) {
-	return function setHook( callback ) {
-		module.hooks[ hookName ].push( callback );
-	};
 }
 
 exportQUnit( QUnit );
